Put the project slides in a separate file
