+++ src/data/projectdata.js
@@ -1,0 +1,18 @@
+import image12 from "./images/events/ros/DSC_0133.JPG";
+
+const slides = [
+    {
+        title: "Project Name 1",
+        description:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis varius lacinia dolor eget dapibus. Fusce interdum risus nunc, nec iaculis sapien tincidunt et. Nullam vel fringilla eros, id mollis ex. Praesent aliquam est sit amet lorem aliquam egestas. Sed consequat, quam a dapibus tempor, purus lorem aliquam leo, sed venenatis.",
+        img: image12,
+    },
+    {
+        title: "Project Name 2",
+        description:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis varius lacinia dolor eget dapibus. Fusce interdum risus nunc, nec iaculis sapien tincidunt et. Nullam vel fringilla eros, id mollis ex. Praesent aliquam est sit amet lorem aliquam egestas. Sed consequat, quam a dapibus tempor, purus lorem aliquam leo, sed venenatis.",
+        img: image12,
+    },
+];
+
+export default slides;
